move some database utilities to common.tools.py
